# testing y api/src/main/java/swagger/api/ActividadesApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.51).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package swagger.api;

import swagger.model.Actividad;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.parameters.RequestBody;

import java.util.List;

import javax.validation.Valid;

import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2023-12-29T13:24:22.602477450Z[GMT]")
@Validated
public interface ActividadesApi {

    @Operation(summary = "Crear una actividad nueva", description = "", tags = {})
    @ApiResponses(value = {
            @ApiResponse(responseCode = "200", description = "Actividad creada con éxito"),

            @ApiResponse(responseCode = "400", description = "Se deben completar los datos de la actividad") })
    @RequestMapping(value = "/actividades/crear", consumes = { "application/json" }, method = RequestMethod.POST)
    ResponseEntity<Void> actividadesCrearPost(
            @Parameter(in = ParameterIn.DEFAULT, description = "Datos de actividad a crear", required = true, schema = @Schema()) @Valid @RequestBody Actividad body);

    @Operation(summary = "Obtener listado de actividades disponibles", description = "", tags = {})
    @ApiResponses(value = {
            @ApiResponse(responseCode = "200", description = "Listado de actividades obtenido correctamente", content = @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Actividad.class)))) })
    @RequestMapping(value = "/actividades/list", produces = { "application/json" }, method = RequestMethod.GET)
    ResponseEntity<List<Actividad>> actividadListGet();

}