# testing y api/target/classes/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Manzanita JR API
  description: API para gestionar actividades y socios en el club Manzanita Jr
  version: 1.0.0
paths:
...
  /actividades/crear:       
    post:
      summary: Crear una actividad nueva (no implementado)
      requestBody:
        description: Datos de actividad a crear
        required: true
        content:
          application/json:
            schema: 
              $ref: '#/components/schemas/Actividad'
      responses: 
        '200':
          description: Actividad creada con éxito
        '400':
          description: Se deben completar los datos de la actividad
...
components:
  schemas:
    Actividad:
      type: object
      properties:
        nombre:
          type: string
          description: Nombre de la actividad
        encargado:
          description: Persona encargada de la actividad
          $ref: '#/components/schemas/Persona'
        inscriptos:
          type: array
          items:
            $ref: '#/components/schemas/Socio'
          description: Lista de socios inscritos en la actividad
        edadMinima:
          type: integer
          description: Edad mínima requerida para participar en la actividad
          example: 25
        cupo:
          type: integer
          description: Cantidad máxima de inscritos permitidos en la actividad
          example: 10
    Persona:
      type: object
      properties:
        nombre:
          type: string
          description: Nombre de la persona
        apellido:
          type: string
          description: Apellido de la persona
        dni:
          type: string
          description: DNI (Documento Nacional de Identidad) de la persona.
          example: 4030303
        edad:
          type: integer
          description: Edad de la persona
          example: 20
    Socio:
      type: object
      properties:
        persona:
          description: Datos personales del socio
          $ref: '#/components/schemas/Persona'
        idSocio:
          type: integer
          description: Identificador único del socio
          example: 12345
        fechaIngreso:
          type: string
          format: date
          description: Fecha de ingreso del socio